Enhance Resilience Against Misconfigured Notification Service Extensions
Improved resilience of the awesome core library to better handle projects that may have misconfigured Notification Service Extensions (NSE) according to the provided documentation. This ensures that even if the NSE is misconfigured, the library can gracefully handle the lack of registered events and other essencial notification informations to continue to operate without causing crashes or major failures.